Moved pages/core.py to pages/core/core.py.
Necessary to refactor classes in core.py to external files.
